+++ opentaps-1.5/opentaps/financials/widget/financials/forms/transactions/TransactionsForms.xml
@@ -568,7 +568,28 @@
         	</field>         	        
         	<field name="area">
         		<display description="${groovy:com.opentaps.financials.Helpers.OperacionDiariaHelper.getEnumName(delegator, mapTransPresup.get(&quot;area&quot;))}"/>
-        	</field> 	        	
+        	</field> 	     
+        	<field name="actividad">
+        		<display description="${groovy:com.opentaps.financials.Helpers.OperacionDiariaHelper.getEnumName(delegator, mapTransPresup.get(&quot;area&quot;))}"/>
+        	</field> 
+        	<field name="subProgramaPresupuestario">
+        		<display description="${groovy:com.opentaps.financials.Helpers.OperacionDiariaHelper.getEnumName(delegator, mapTransPresup.get(&quot;subProgramaPresupuestario&quot;))}"/>
+        	</field>
+        	<field name="programaPresupuestario">
+        		<display description="${groovy:com.opentaps.financials.Helpers.OperacionDiariaHelper.getEnumName(delegator, mapTransPresup.get(&quot;programaPresupuestario&quot;))}"/>
+        	</field>        
+        	<field name="programaPlan">
+        		<display description="${groovy:com.opentaps.financials.Helpers.OperacionDiariaHelper.getEnumName(delegator, mapTransPresup.get(&quot;programaPlan&quot;))}"/>
+        	</field>
+        	<field name="idPago">
+        		<display description="${groovy:com.opentaps.financials.Helpers.OperacionDiariaHelper.getPaymenthMethodName(delegator, mapTransPresup.get(&quot;idPago&quot;))}"/>
+        	</field>        	   
+        	<field name="idProductoD">
+        		<display description="${groovy:com.opentaps.financials.Helpers.OperacionDiariaHelper.getProductName(delegator, mapTransPresup.get(&quot;idProductoD&quot;))}"/>
+        	</field>
+        	<field name="idProductoH">
+        		<display description="${groovy:com.opentaps.financials.Helpers.OperacionDiariaHelper.getProductName(delegator, mapTransPresup.get(&quot;idProductoH&quot;))}"/>
+        	</field> 		        	
     </form>
     
     <!--  Esme Presupuesto formas  -->    
@@ -623,7 +644,7 @@
         	</field>
        		<field name="idTipoDoc">
             	<display description="${groovy:com.opentaps.financials.Helpers.OperacionDiariaHelper.getTipoDocName(delegator, acctgTransPresupuestal.get(&quot;idTipoDoc&quot;))}" />
-        	</field>        
+        	</field>     
     </form>
     
 </forms>
